Add ReplicateEnabled check, remove file limit in ProcessFiles
Introduced a new environment variable "REPLICATE_ENABLED" to manage the execution of the replicating process. The ReplicateEnabled() method in the EnvHelper class checks for this variable. Updated LlmFileUploaderHandler.cs to check whether ReplicateEnabled is true before sending a prediction. Removed the temporary file limit in the ProcessFiles method to process all files.