"Add detection for inline comments in CommentChecker.cs"
A new regular expression was added in the CommentChecker script to detect inline comments in files. This addition will help to differentiate between regular and inline comments for a more accurate comment count. As a result, the number of false positives where a comment might be missed can be reduced. Additional lines of codes were also added for the test cases, for a more rigorous testing of the updated script.