Refactor CommentChecker class for efficiency
Adjusted `CommentChecker` class to store `ZipArchiveEntry` file and different types of comments as private variables rather than passing them as parameters between functions. This change was made to simplify the function calls throughout the class and decrease the number of parameters being passed around, thus improving the readability of the code and optimizing it for efficiency.