Add counter to log file upload progress
It was difficult to track the progress of file uploading in the system logs. To solve this, a counter has been added that increments with each file processed by the uploader. The count of the current file being processed and total number of files will now be displayed in the logs.