Refactor diff calculation and rename properties for clarity
Renamed the properties 'old' and 'new' in SetupDiffDataFlat.js and the corresponding references in other files to 'oldValue' and 'newValue' respectively. Additionally, refactored the diff calculation in +page.svelte to encapsulate each string character into its own object. This was done to handle selection of specific lines in diffs, instead of selecting entire older or newer versions. The commit will help to provide a clearer diff view and improve user interaction.