Add debugging logs to review diff merging process
Two console.log statements have been added to the inner check of the diff merging process in the '+page.svelte' file. This change was made to debug and verify the accuracy of the diff merging operation. By logging the merged diffs and the lineObject, developers can now directly monitor and verify whether the correct data is being combined.