Add JSZip to create download files in app
Added JSZip dependency to allow file content zipping before triggering download in the client. It eliminates the need of downloading multiple files as each change will be saved into a single .zip file. Refactored the way data strings are concatenated to avoid unnecessary newlines at the start of the string, which had a negative impact on the formatting of exported files.